site-form-works/avia_scanner/requirements.txt
Django==2.1.1
django_select2
django-bootstrap-datepicker-plus

